working multiple masters, multiple learners with learners on workers also

+++ source/Learners/Learner.h
@@ -103,7 +103,7 @@ public:
   long locDataSetSize() const {
     return data->readNData();
   }
-  unsigned nSeqsEval() const {
+  long nSeqsEval() const {
     return data->readNSeenSeq_loc();
   }
   long nGradSteps() const {
